feat: rename kafka -> cdc dir, remove redundant topic array
