+++ packages/indexer/src/cdc/topics/indexer-fill-events.ts
@@ -2,11 +2,11 @@
 
 import { logger } from "@/common/logger";
 import { redisWebsocketPublisher } from "@/common/redis";
-import { KafkaTopicHandler } from "kafka";
+import { KafkaTopicHandler } from "cdc";
 
 // Create a class implementing KafkaEventHandler for each event type
-export class IndexerOrderEventsHandler implements KafkaTopicHandler {
-  topicName = "indexer.public.order_events";
+export class IndexerFillEventsHandler implements KafkaTopicHandler {
+  topicName = "indexer.public.fill_events_2";
 
   async handle(payload: any): Promise<void> {
     // eslint-disable-next-line no-console
@@ -37,44 +37,26 @@ export class IndexerOrderEventsHandler implements KafkaTopicHandler {
       return;
     }
 
-    if (payload.after.kind === "new-order") {
-      // trigger ask.created event
+    await redisWebsocketPublisher.publish(
+      "events",
+      JSON.stringify({
+        event: "sell.created.v2",
+        tags: {},
+        data: payload.after,
+      })
+    );
+  }
 
-      await redisWebsocketPublisher.publish(
-        "events",
-        JSON.stringify({
-          event: "ask.created.v2",
-          tags: {
-            contract: payload.after.contract,
-          },
-          data: payload.after,
-        })
-      );
-
+  async handleUpdate(payload: any): Promise<void> {
+    if (!payload.after) {
       return;
     }
 
     await redisWebsocketPublisher.publish(
       "events",
       JSON.stringify({
-        event: "ask.updated.v2",
-        tags: {
-          contract: payload.after.contract,
-        },
-        data: payload.after,
-      })
-    );
-    // all other cases, trigger ask.updated event
-  }
-
-  async handleUpdate(payload: any): Promise<void> {
-    await redisWebsocketPublisher.publish(
-      "events",
-      JSON.stringify({
-        event: "ask.updated.v2",
-        tags: {
-          contract: payload.after.contract,
-        },
+        event: "sell.updated.v2",
+        tags: {},
         data: payload.after,
       })
     );
