+++ packages/indexer/src/cdc/topics/indexer-order-events.ts
@@ -2,11 +2,11 @@
 
 import { logger } from "@/common/logger";
 import { redisWebsocketPublisher } from "@/common/redis";
-import { KafkaTopicHandler } from "kafka";
+import { KafkaTopicHandler } from "cdc";
 
 // Create a class implementing KafkaEventHandler for each event type
-export class IndexerTransferEventsHandler implements KafkaTopicHandler {
-  topicName = "indexer.public.ft_transfer_events";
+export class IndexerOrderEventsHandler implements KafkaTopicHandler {
+  topicName = "indexer.public.order_events";
 
   async handle(payload: any): Promise<void> {
     // eslint-disable-next-line no-console
@@ -37,27 +37,44 @@ export class IndexerTransferEventsHandler implements KafkaTopicHandler {
       return;
     }
 
-    await redisWebsocketPublisher.publish(
-      "events",
-      JSON.stringify({
-        event: "transfer.created.v2",
-        tags: {},
-        data: payload.after,
-      })
-    );
-  }
+    if (payload.after.kind === "new-order") {
+      // trigger ask.created event
 
-  async handleUpdate(payload: any): Promise<void> {
-    // probably do nothing here
-    if (!payload.after) {
+      await redisWebsocketPublisher.publish(
+        "events",
+        JSON.stringify({
+          event: "ask.created.v2",
+          tags: {
+            contract: payload.after.contract,
+          },
+          data: payload.after,
+        })
+      );
+
       return;
     }
 
     await redisWebsocketPublisher.publish(
       "events",
       JSON.stringify({
-        event: "transfer.updated.v2",
-        tags: {},
+        event: "ask.updated.v2",
+        tags: {
+          contract: payload.after.contract,
+        },
+        data: payload.after,
+      })
+    );
+    // all other cases, trigger ask.updated event
+  }
+
+  async handleUpdate(payload: any): Promise<void> {
+    await redisWebsocketPublisher.publish(
+      "events",
+      JSON.stringify({
+        event: "ask.updated.v2",
+        tags: {
+          contract: payload.after.contract,
+        },
         data: payload.after,
       })
     );
